Split out reducing and mutants to their own headers
